feat(phone/settings): call volume slider in settings (#484)
You can now adjust the user's call volume by changing the slider in the Settings App.
