+++ phone/src/apps/settings/state/customWallpaper.state.ts
@@ -1,0 +1,8 @@
+import { atom, useRecoilState } from 'recoil';
+
+const customWallpaperState = atom({
+  key: 'customWallpaperState',
+  default: false,
+});
+
+export const useCustomWallpaperModal = () => useRecoilState(customWallpaperState);
