+++ phone/src/os/apps/hooks/useApps.tsx
@@ -5,11 +5,7 @@ import { APPS, IApp } from '../config/apps';
 import { SvgIconComponent } from '@mui/icons-material';
 import { useTheme } from '@mui/material';
 import { useSettingsValue } from '../../../apps/settings/hooks/useSettings';
-
-export interface IconSetObject {
-  custom: boolean;
-  name: string;
-}
+import { IconSetObject } from '@typings/settings';
 
 export const useApps = () => {
   const { icons } = useNotifications();
